requirements.txt
Django>=4.0
djangorestframework
django-extensions
Werkzeug
pyOpenSSL
psycopg2-binary
python-dotenv
drf-spectacular
drf-spectacular-sidecar
django-cors-headers>=4.0.0
google-auth
requests
PyPDF2>=3.0.0
djangorestframework-simplejwt
langid
openai>=0.27.0
tiktoken>=0.3.0
pikepdf>=5.0.0
django-cors-headers
django-filter